# feat: implement high-performance BLAKE3 Content-Addressable Storage core