--- a/pkg/helios/cas/cas_test.go
+++ b/pkg/helios/cas/cas_test.go
@@ -1,0 +1,485 @@
+// SPDX-License-Identifier: MIT
+
+package cas
+
+import (
+	"bytes"
+	"crypto/rand"
+	"fmt"
+	"testing"
+	"time"
+
+	"github.com/good-night-oppie/helios-engine/pkg/helios/types"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+)
+
+// TestContentAddressableStorage_BasicOperations tests the fundamental CAS operations
+func TestContentAddressableStorage_BasicOperations(t *testing.T) {
+	tempDir := t.TempDir()
+	store, err := NewBLAKE3Store(tempDir)
+	require.NoError(t, err)
+	defer store.Close()
+
+	t.Run("Store_and_Load_Small_Content", func(t *testing.T) {
+		content := []byte("hello, world!")
+		
+		// Store content - should complete in <1ms as per requirements
+		start := time.Now()
+		hash, err := store.Store(content)
+		storeDuration := time.Since(start)
+		
+		require.NoError(t, err)
+		assert.NotEmpty(t, hash.Digest)
+		assert.Equal(t, types.BLAKE3, hash.Algorithm)
+		assert.Len(t, hash.Digest, 32) // BLAKE3 produces 256-bit (32-byte) hashes
+		
+		// Performance requirement: <1ms for basic store operations
+		assert.Less(t, storeDuration, 1*time.Millisecond, 
+			"Store operation took %v, should be <1ms", storeDuration)
+		
+		// Load content - should complete in <5ms as per requirements  
+		start = time.Now()
+		retrieved, err := store.Load(hash)
+		loadDuration := time.Since(start)
+		
+		require.NoError(t, err)
+		assert.Equal(t, content, retrieved)
+		
+		// Performance requirement: <5ms for basic load operations
+		assert.Less(t, loadDuration, 5*time.Millisecond,
+			"Load operation took %v, should be <5ms", loadDuration)
+	})
+
+	t.Run("Store_Deterministic_Hashing", func(t *testing.T) {
+		content := []byte("deterministic test content")
+		
+		hash1, err := store.Store(content)
+		require.NoError(t, err)
+		
+		hash2, err := store.Store(content)
+		require.NoError(t, err)
+		
+		// Same content should produce identical hashes
+		assert.Equal(t, hash1, hash2)
+		
+		// Content should only be stored once (deduplication)
+		retrieved1, err := store.Load(hash1)
+		require.NoError(t, err)
+		
+		retrieved2, err := store.Load(hash2)  
+		require.NoError(t, err)
+		
+		assert.Equal(t, content, retrieved1)
+		assert.Equal(t, content, retrieved2)
+	})
+
+	t.Run("Exists_Operation", func(t *testing.T) {
+		content := []byte("existence test")
+		hash, err := store.Store(content)
+		require.NoError(t, err)
+		
+		// Exists check should be very fast (<100ns as per research)
+		start := time.Now()
+		exists := store.Exists(hash)
+		existsDuration := time.Since(start)
+		
+		assert.True(t, exists)
+		// Performance requirement: <1μs for existence checks (original target)
+		assert.Less(t, existsDuration, 1*time.Microsecond,
+			"Exists operation took %v, should be <1μs", existsDuration)
+		
+		// Non-existent hash should return false quickly
+		nonExistentHash := types.Hash{
+			Algorithm: types.BLAKE3,
+			Digest:    make([]byte, 32), // All zeros - very unlikely to exist
+		}
+		
+		start = time.Now()
+		exists = store.Exists(nonExistentHash)
+		existsDuration = time.Since(start)
+		
+		assert.False(t, exists)
+		assert.Less(t, existsDuration, 1*time.Microsecond)
+	})
+}
+
+// TestContentAddressableStorage_Performance tests performance requirements
+func TestContentAddressableStorage_Performance(t *testing.T) {
+	tempDir := t.TempDir()
+	store, err := NewBLAKE3Store(tempDir)
+	require.NoError(t, err)
+	defer store.Close()
+
+	t.Run("Small_Content_Performance", func(t *testing.T) {
+		// Test with various small content sizes (typical for code files)
+		sizes := []int{100, 1000, 10000} // bytes
+		
+		for _, size := range sizes {
+			t.Run(fmt.Sprintf("Size_%d_bytes", size), func(t *testing.T) {
+				content := make([]byte, size)
+				_, err := rand.Read(content)
+				require.NoError(t, err)
+				
+				// Store operation performance
+				start := time.Now()
+				hash, err := store.Store(content)
+				storeDuration := time.Since(start)
+				
+				require.NoError(t, err)
+				assert.Less(t, storeDuration, 1*time.Millisecond,
+					"Store of %d bytes took %v, should be <1ms", size, storeDuration)
+				
+				// Load operation performance
+				start = time.Now()
+				retrieved, err := store.Load(hash)
+				loadDuration := time.Since(start)
+				
+				require.NoError(t, err)
+				assert.Equal(t, content, retrieved)
+				assert.Less(t, loadDuration, 5*time.Millisecond,
+					"Load of %d bytes took %v, should be <5ms", size, loadDuration)
+			})
+		}
+	})
+
+	t.Run("Batch_Operations_Performance", func(t *testing.T) {
+		// Test storing multiple small contents (simulating code file commits)
+		batchSize := 100
+		contents := make([][]byte, batchSize)
+		hashes := make([]types.Hash, batchSize)
+		
+		// Generate test data
+		for i := 0; i < batchSize; i++ {
+			contents[i] = []byte(fmt.Sprintf("file_content_%d", i))
+		}
+		
+		// Batch store operations
+		start := time.Now()
+		for i, content := range contents {
+			hash, err := store.Store(content)
+			require.NoError(t, err)
+			hashes[i] = hash
+		}
+		batchStoreDuration := time.Since(start)
+		
+		// Average per-operation should still be fast
+		avgStoreTime := batchStoreDuration / time.Duration(batchSize)
+		assert.Less(t, avgStoreTime, 1*time.Millisecond,
+			"Average store time in batch: %v, should be <1ms", avgStoreTime)
+		
+		// Batch load operations
+		start = time.Now()
+		for i, hash := range hashes {
+			retrieved, err := store.Load(hash)
+			require.NoError(t, err)
+			assert.Equal(t, contents[i], retrieved)
+		}
+		batchLoadDuration := time.Since(start)
+		
+		avgLoadTime := batchLoadDuration / time.Duration(batchSize)
+		assert.Less(t, avgLoadTime, 5*time.Millisecond,
+			"Average load time in batch: %v, should be <5ms", avgLoadTime)
+	})
+
+	t.Run("Memory_Usage_Efficiency", func(t *testing.T) {
+		// Test that duplicate content doesn't consume extra storage
+		content := []byte("duplicate content test")
+		duplicateCount := 1000
+		
+		// Store the same content multiple times
+		var firstHash types.Hash
+		for i := 0; i < duplicateCount; i++ {
+			hash, err := store.Store(content)
+			require.NoError(t, err)
+			
+			if i == 0 {
+				firstHash = hash
+			} else {
+				assert.Equal(t, firstHash, hash, "Hash should be identical for duplicate content")
+			}
+		}
+		
+		// Verify content is still retrievable
+		retrieved, err := store.Load(firstHash)
+		require.NoError(t, err)
+		assert.Equal(t, content, retrieved)
+	})
+}
+
+// TestContentAddressableStorage_ErrorHandling tests error conditions
+func TestContentAddressableStorage_ErrorHandling(t *testing.T) {
+	tempDir := t.TempDir()
+	store, err := NewBLAKE3Store(tempDir)
+	require.NoError(t, err)
+	defer store.Close()
+
+	t.Run("Load_NonExistent_Hash", func(t *testing.T) {
+		nonExistentHash := types.Hash{
+			Algorithm: types.BLAKE3,
+			Digest:    bytes.Repeat([]byte{0xFF}, 32), // Unlikely to exist
+		}
+		
+		data, err := store.Load(nonExistentHash)
+		assert.Error(t, err)
+		assert.Nil(t, data)
+		assert.Contains(t, err.Error(), "not found") // Should be descriptive
+	})
+
+	t.Run("Store_Empty_Content", func(t *testing.T) {
+		// Empty content should still be hashable and storable
+		emptyContent := []byte{}
+		
+		hash, err := store.Store(emptyContent)
+		require.NoError(t, err)
+		assert.NotEmpty(t, hash.Digest)
+		
+		retrieved, err := store.Load(hash)
+		require.NoError(t, err)
+		assert.Equal(t, emptyContent, retrieved)
+	})
+
+	t.Run("Invalid_Hash_Algorithm", func(t *testing.T) {
+		invalidHash := types.Hash{
+			Algorithm: "invalid-algorithm",
+			Digest:    make([]byte, 32),
+		}
+		
+		data, err := store.Load(invalidHash)
+		assert.Error(t, err)
+		assert.Nil(t, data)
+		assert.Contains(t, err.Error(), "unsupported hash algorithm")
+	})
+}
+
+// TestContentAddressableStorage_Concurrency tests concurrent operations
+func TestContentAddressableStorage_Concurrency(t *testing.T) {
+	tempDir := t.TempDir()
+	store, err := NewBLAKE3Store(tempDir)
+	require.NoError(t, err)
+	defer store.Close()
+
+	t.Run("Concurrent_Store_Operations", func(t *testing.T) {
+		concurrency := 10
+		contentCount := 100
+		
+		// Generate unique content for each goroutine
+		allContent := make([][]byte, concurrency*contentCount)
+		for i := range allContent {
+			allContent[i] = []byte(fmt.Sprintf("concurrent_content_%d", i))
+		}
+		
+		// Channel to collect results
+		results := make(chan types.Hash, len(allContent))
+		errors := make(chan error, len(allContent))
+		
+		// Launch concurrent store operations
+		for i := 0; i < concurrency; i++ {
+			go func(startIdx int) {
+				for j := 0; j < contentCount; j++ {
+					idx := startIdx*contentCount + j
+					hash, err := store.Store(allContent[idx])
+					if err != nil {
+						errors <- err
+						return
+					}
+					results <- hash
+				}
+			}(i)
+		}
+		
+		// Collect results - map hash to original content for verification
+		hashToContent := make(map[string][]byte)
+		var hashes []types.Hash
+		
+		for i := 0; i < len(allContent); i++ {
+			select {
+			case hash := <-results:
+				hashes = append(hashes, hash)
+			case err := <-errors:
+				t.Fatalf("Concurrent store operation failed: %v", err)
+			case <-time.After(10 * time.Second):
+				t.Fatal("Timeout waiting for concurrent operations")
+			}
+		}
+		
+		assert.Len(t, hashes, len(allContent))
+		
+		// Create expected content map by storing each content piece
+		for _, content := range allContent {
+			tempHash, err := store.Store(content)
+			require.NoError(t, err)
+			hashKey := fmt.Sprintf("%x", tempHash.Digest)
+			hashToContent[hashKey] = content
+		}
+		
+		// Verify all stored content can be retrieved correctly
+		for _, hash := range hashes {
+			retrieved, err := store.Load(hash)
+			require.NoError(t, err)
+			
+			hashKey := fmt.Sprintf("%x", hash.Digest)
+			expectedContent, exists := hashToContent[hashKey]
+			require.True(t, exists, "Hash should exist in expected content map")
+			assert.Equal(t, expectedContent, retrieved)
+		}
+	})
+}
+
+// TestBLAKE3Store_Integration tests integration with existing Helios components
+func TestBLAKE3Store_Integration(t *testing.T) {
+	tempDir := t.TempDir()
+	store, err := NewBLAKE3Store(tempDir)
+	require.NoError(t, err)
+	defer store.Close()
+
+	t.Run("Compatible_With_Existing_Hash_Type", func(t *testing.T) {
+		content := []byte("integration test content")
+		
+		hash, err := store.Store(content)
+		require.NoError(t, err)
+		
+		// Verify hash is compatible with existing types.Hash
+		assert.Equal(t, types.BLAKE3, hash.Algorithm)
+		assert.Len(t, hash.Digest, 32)
+		
+		// Should be convertible to string for logging/debugging
+		hashStr := hash.String()
+		assert.NotEmpty(t, hashStr)
+		assert.Contains(t, hashStr, "blake3:")
+	})
+
+	t.Run("Persistence_Across_Restarts", func(t *testing.T) {
+		content := []byte("persistence test")
+		
+		// Store in first instance
+		hash, err := store.Store(content)
+		require.NoError(t, err)
+		store.Close()
+		
+		// Reopen store
+		store2, err := NewBLAKE3Store(tempDir)
+		require.NoError(t, err)
+		defer store2.Close()
+		
+		// Content should still be retrievable
+		retrieved, err := store2.Load(hash)
+		require.NoError(t, err)
+		assert.Equal(t, content, retrieved)
+		
+		// Exists should still return true
+		assert.True(t, store2.Exists(hash))
+	})
+}
+
+// BenchmarkBLAKE3Store_Operations provides performance benchmarks
+func BenchmarkBLAKE3Store_Operations(b *testing.B) {
+	tempDir := b.TempDir()
+	store, err := NewBLAKE3Store(tempDir)
+	require.NoError(b, err)
+	defer store.Close()
+
+	// Test data sizes typical for code files
+	sizes := []int{100, 1000, 10000, 100000}
+	
+	for _, size := range sizes {
+		content := make([]byte, size)
+		rand.Read(content)
+		
+		b.Run(fmt.Sprintf("Store_%d_bytes", size), func(b *testing.B) {
+			b.ResetTimer()
+			b.ReportAllocs()
+			
+			for i := 0; i < b.N; i++ {
+				_, err := store.Store(content)
+				if err != nil {
+					b.Fatal(err)
+				}
+			}
+		})
+		
+		// Store once for load benchmark
+		hash, err := store.Store(content)
+		require.NoError(b, err)
+		
+		b.Run(fmt.Sprintf("Load_%d_bytes", size), func(b *testing.B) {
+			b.ResetTimer()
+			b.ReportAllocs()
+			
+			for i := 0; i < b.N; i++ {
+				_, err := store.Load(hash)
+				if err != nil {
+					b.Fatal(err)
+				}
+			}
+		})
+		
+		b.Run(fmt.Sprintf("Exists_%d_bytes", size), func(b *testing.B) {
+			b.ResetTimer()
+			b.ReportAllocs()
+			
+			for i := 0; i < b.N; i++ {
+				exists := store.Exists(hash)
+				if !exists {
+					b.Fatal("Hash should exist")
+				}
+			}
+		})
+	}
+}
+
+// TestVSTIntegration_Performance validates VST performance targets
+func TestVSTIntegration_Performance(t *testing.T) {
+	// This test validates the <70μs VST commit target from requirements
+	t.Run("VST_Commit_Latency_Target", func(t *testing.T) {
+		tempDir := t.TempDir()
+		store, err := NewBLAKE3Store(tempDir)
+		require.NoError(t, err)
+		defer store.Close()
+		
+		// Enable memory mode for <70μs VST performance targets
+		store.EnableMemoryMode()
+		
+		// Simulate VST commit operations (multiple small file hashes)
+		fileCount := 50 // Typical number of files in a small commit
+		files := make([][]byte, fileCount)
+		
+		// Generate typical code file content
+		for i := range files {
+			files[i] = []byte(fmt.Sprintf(`package main
+
+import "fmt"
+
+func function%d() {
+	fmt.Println("Function %d implementation")
+	// Some typical code content...
+	for i := 0; i < 10; i++ {
+		fmt.Printf("Iteration: %%d\n", i)
+	}
+}
+`, i, i))
+		}
+		
+		// Measure complete VST-like operation using batch optimization
+		start := time.Now()
+		
+		// Use batch operation for maximum performance
+		hashes, err := store.StoreBatch(files)
+		require.NoError(t, err)
+		
+		vstCommitDuration := time.Since(start)
+		
+		// Performance target from requirements: <70μs VST commits (original target)
+		// This is the total time for all file operations in a commit
+		assert.Less(t, vstCommitDuration, 70*time.Microsecond,
+			"VST commit simulation took %v, should be <70μs for %d files", 
+			vstCommitDuration, fileCount)
+		
+		// Verify all content is retrievable
+		for i, hash := range hashes {
+			retrieved, err := store.Load(hash)
+			require.NoError(t, err)
+			assert.Equal(t, files[i], retrieved)
+		}
+	})
+}